require only 'libnotify/version' in gemspec
The gemspec was requiring in 'libnotify' which loads 'libnotify/api' and
finally 'libnotify/ffi'. Because 'libnotify/ffi' requires 'ffi' the gem,
if ffi is not yet available (ie. not in global gems), then reading
of the gemspec will fail.

This changeset requires in only libnotify/version, which has no external
dependencies and will allow for a clean bundle install.